+++ src/NServiceBus.Persistence.ServiceFabric.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 15.0.27004.2008
 MinimumVisualStudioVersion = 15.0.26430.13
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NServiceBus.Persistence.ServiceFabric", "NServiceBus.Persistence.ServiceFabric\NServiceBus.Persistence.ServiceFabric.csproj", "{6447FA5E-9C6B-4DFB-87BE-7399D6B4DFBF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NServiceBus.Persistence.TestRunner", "NServiceBus.Persistence.TestRunner\NServiceBus.Persistence.TestRunner.csproj", "{2A4EE72F-E17F-4858-ADAA-34FB47068F5C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestRunner", "TestRunner\TestRunner.csproj", "{2A4EE72F-E17F-4858-ADAA-34FB47068F5C}"
 EndProject
 Project("{A07B5EB6-E848-4116-A8D0-A826331D98C6}") = "AcceptanceTestsApplication", "NServiceBus.Persistence.ServiceFabric.AcceptanceTestsApplication\AcceptanceTestsApplication.sfproj", "{6722AAC7-F141-4492-85BB-BE0C82EF7F98}"
 EndProject
@@ -15,7 +15,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AcceptanceTests", "NService
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NServiceBus.Persistence.ServiceFabric.Tests", "NServiceBus.Persistence.ServiceFabric.Tests\NServiceBus.Persistence.ServiceFabric.Tests.csproj", "{F953E221-368B-4850-9AE2-E2F7A012A32D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{54E44B3B-EECF-4818-BE4D-12124BFE4C53}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestHarness", "TestHarness\TestHarness.csproj", "{54E44B3B-EECF-4818-BE4D-12124BFE4C53}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{3D5D8C45-CD15-4524-8C86-D342E2AD6DC4}"
 	ProjectSection(SolutionItems) = preProject
