+++ src/TestHarness/AcceptanceTests.cs
@@ -1,4 +1,4 @@
-namespace Tests
+namespace TestHarness
 {
     public class AcceptanceTests : R<AcceptanceTests>
     {
