+++ src/TestHarness/TestHarness.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\NServiceBus.Persistence.TestRunner\NServiceBus.Persistence.TestRunner.csproj" />
+    <ProjectReference Include="..\TestRunner\TestRunner.csproj" />
   </ItemGroup>
 
   <ItemGroup>
