+++ src/TestHarness/Tests.cs
@@ -1,4 +1,4 @@
-namespace Tests
+namespace TestHarness
 {
     public class Tests : R<Tests>
     {
